refactor: split chapter nav internals
