+++ frontend/src/components/editor/nav/ChapterActionMenu.vue
@@ -1,18 +1,11 @@
 <script lang="ts" setup>
-import { flagUrl } from "@/utils/media";
+import type { ChapterStartOption } from "./types";
 import { useI18n } from "vue-i18n";
-import { computed } from "vue";
+import ChapterStartSelect from "./ChapterStartSelect.vue";
 import {
   symOutlinedCallMerge,
   symOutlinedCallSplit,
 } from "@quasar/extras/material-symbols-outlined";
-
-type StartOption = {
-  label: string;
-  value: number;
-  countryCode: string;
-  countryLabel: string;
-};
 
 const { t } = useI18n();
 
@@ -20,15 +13,9 @@ const props = defineProps<{
   canDelete?: boolean;
   canSplit?: boolean;
   mergeTarget?: "previous" | "next";
-  startOptions?: StartOption[];
+  startOptions?: ChapterStartOption[];
   startStepId?: number | null;
 }>();
-
-const selectedStartOption = computed(
-  () =>
-    props.startOptions?.find((option) => option.value === props.startStepId) ??
-    null,
-);
 
 defineEmits<{
   splitChapter: [];
@@ -41,56 +28,11 @@ defineEmits<{
   <q-list dense class="chapter-action-menu">
     <q-item v-if="startOptions?.length" class="chapter-start-item">
       <q-item-section>
-        <q-select
-          :model-value="startStepId"
+        <ChapterStartSelect
+          :model-value="props.startStepId"
           :options="startOptions"
-          :label="t('chapters.startsAt')"
-          class="chapter-start-select"
-          popup-content-class="chapter-start-popup"
-          dense
-          borderless
-          emit-value
-          map-options
-          options-dense
-          @update:model-value="$emit('adjustBoundary', Number($event))"
-        >
-          <template #selected>
-            <div
-              v-if="selectedStartOption"
-              class="chapter-start-selected"
-              dir="ltr"
-            >
-              <img
-                :src="flagUrl(selectedStartOption.countryCode)"
-                :alt="selectedStartOption.countryLabel"
-                class="chapter-start-flag"
-              />
-              <span class="chapter-start-label" dir="auto">
-                {{ selectedStartOption.label }}
-              </span>
-            </div>
-          </template>
-          <template #option="scope">
-            <q-item
-              v-bind="scope.itemProps"
-              class="chapter-start-option"
-              dir="ltr"
-            >
-              <q-item-section side class="chapter-start-option-flag">
-                <img
-                  :src="flagUrl(scope.opt.countryCode)"
-                  :alt="scope.opt.countryLabel"
-                  class="chapter-start-flag"
-                />
-              </q-item-section>
-              <q-item-section class="chapter-start-option-text">
-                <span class="chapter-start-label" dir="auto">
-                  {{ scope.opt.label }}
-                </span>
-              </q-item-section>
-            </q-item>
-          </template>
-        </q-select>
+          @update:model-value="$emit('adjustBoundary', $event)"
+        />
       </q-item-section>
     </q-item>
     <q-separator v-if="startOptions?.length" class="chapter-menu-separator" />
@@ -141,74 +83,6 @@ defineEmits<{
   min-height: 3.25rem;
   padding: 0 var(--gap-sm);
   border-radius: var(--radius-sm);
-
-  :deep(.q-field__control) {
-    min-height: 2.75rem;
-    padding-inline: 0;
-  }
-
-  :deep(.q-field__label) {
-    color: var(--text-muted);
-    font-size: var(--type-xs);
-    font-weight: 650;
-  }
-
-  :deep(.q-field__native) {
-    color: var(--text-bright);
-    font-size: var(--type-sm);
-    font-weight: 600;
-  }
-}
-
-.chapter-start-select {
-  min-width: 0;
-}
-
-.chapter-start-selected,
-:global(.chapter-start-option) {
-  display: flex;
-  align-items: center;
-  gap: var(--gap-sm);
-  min-width: 0;
-  text-align: left;
-}
-
-.chapter-start-selected {
-  width: 100%;
-}
-
-.chapter-start-flag {
-  width: 0.875rem;
-  height: 0.625rem;
-  flex: 0 0 auto;
-  border-radius: var(--radius-xs);
-}
-
-.chapter-start-label {
-  min-width: 0;
-  overflow: hidden;
-  text-align: left;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-:global(.chapter-start-popup) {
-  direction: ltr;
-  text-align: left;
-}
-
-:global(.chapter-start-option) {
-  min-height: 2.25rem;
-  padding-inline: var(--gap-sm);
-}
-
-:global(.chapter-start-option-flag) {
-  min-width: unset;
-  padding-inline-end: 0;
-}
-
-:global(.chapter-start-option-text) {
-  min-width: 0;
 }
 
 .chapter-menu-separator {
